Migrate to Bun, Biome, structure changes, CI/CD update
- Migrate from PNPM to Bun (package manager)
- Migrate from Node to Bun (runtime) for /scripts
- Migrate from Prettier/ESLint to Biome
- Migrate Tests from Jest to Bun Test Runner
- Move tests closer to each tested file, also splitting them
- Update GH Action to support NPM's Trusted publishing, Bun and update used actions' versions
- Update README, CONTRIBUTING and LICENSE year
