--- .eslintrc.js
@@ -1,20 +1,0 @@
-module.exports = {
-  env: {
-    browser: true,
-    es2021: true,
-    node: true,
-  },
-  extends: ['airbnb-base', 'airbnb-typescript/base'],
-  parser: '@typescript-eslint/parser',
-  parserOptions: {
-    ecmaVersion: 12,
-    sourceType: 'module',
-    project: './tsconfig.json',
-  },
-  plugins: ['@typescript-eslint'],
-  rules: {
-    semi: 'off',
-    '@typescript-eslint/semi': 'off',
-  },
-  ignorePatterns: ['.eslintrc.js'],
-}
