+++ LICENSE
@@ -632,7 +632,7 @@ state the exclusion of warranty; and each file should have at least
 the "copyright" line and a pointer to where the full notice is found.
 
     dr-utils: Utilities relevant to the Dominican Republic
-    Copyright (C) 2021  Miguel Araujo
+    Copyright (C) 2026  Miguel Araujo
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
