--- jest.config.js
@@ -1,11 +1,0 @@
-module.exports = {
-  roots: ['<rootDir>'],
-  testMatch: [
-    '**/tests/**/*.+(ts|tsx|js)',
-    '**/__tests__/**/*.+(ts|tsx|js)',
-    '**/?(*.)+(spec|test).+(ts|tsx|js)',
-  ],
-  transform: {
-    '^.+\\.(ts|tsx)$': 'ts-jest',
-  },
-}
