+++ src/formatters/phone.formatter.ts
@@ -1,6 +1,6 @@
 /*
 dr-utils: Utilities relevant to the Dominican Republic
-Copyright (C) 2021  Miguel Araujo
+Copyright (C) 2026  Miguel Araujo
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -17,34 +17,30 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
 import {
-  PHONE_NUMBER_EXTRACT_REGEX,
-  PHONE_NUMBER_VALID_REGEX,
+	PHONE_NUMBER_EXTRACT_REGEX,
+	PHONE_NUMBER_VALID_REGEX,
 } from '../constants/phone-number.regex'
 
 /**
- * @param phoneNumber {string} the phone number
- * @param international {boolean} if the output is national
- * format (`false`) or international (`true`)
- * @returns {string} the formatted phone number
+ * @param phoneNumber the phone number
+ * @param international if the output should be the international format (E.164) or not (domestic)
+ * @returns the formatted phone number
  */
-export function formatPhoneNumber(
-  phoneNumber: string,
-  international: boolean = false,
-): string {
-  const digitsOnly = phoneNumber.replace(PHONE_NUMBER_EXTRACT_REGEX, '')
-  const matches = PHONE_NUMBER_VALID_REGEX.exec(digitsOnly)
+export function formatPhoneNumber(phoneNumber: string, international: boolean = false): string {
+	const digitsOnly = phoneNumber.replace(PHONE_NUMBER_EXTRACT_REGEX, '')
+	const matches = digitsOnly.match(PHONE_NUMBER_VALID_REGEX)
 
-  if (!matches) throw new Error(`Phone Number "${phoneNumber}" is not valid`)
+	if (!matches) throw new Error(`Phone Number "${phoneNumber}" is not valid`)
 
-  const areaCode = matches[1]
-  const prefix = matches[2]
-  const line = matches[3]
+	const areaCode = matches[1]
+	const prefix = matches[2]
+	const line = matches[3]
 
-  if (!international) {
-    return `(${areaCode}) ${prefix}-${line}`
-  }
+	if (!international) {
+		return `(${areaCode}) ${prefix}-${line}`
+	}
 
-  return `+1 ${areaCode} ${prefix} ${line}`
+	return `+1${areaCode}${prefix}${line}`
 }
 
 export default formatPhoneNumber
