+++ src/validators/phone.validator.ts
@@ -1,6 +1,6 @@
 /*
 dr-utils: Utilities relevant to the Dominican Republic
-Copyright (C) 2021  Miguel Araujo
+Copyright (C) 2026  Miguel Araujo
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -17,8 +17,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
 import {
-  PHONE_NUMBER_EXTRACT_REGEX,
-  PHONE_NUMBER_VALID_REGEX,
+	PHONE_NUMBER_EXTRACT_REGEX,
+	PHONE_NUMBER_VALID_REGEX,
 } from '../constants/phone-number.regex'
 
 /**
@@ -26,9 +26,9 @@ import {
  * @returns {boolean} `true` if the number is valid, `false` otherwise
  */
 export function validatePhoneNumber(phoneNumber: string): boolean {
-  const stripped = phoneNumber.replace(PHONE_NUMBER_EXTRACT_REGEX, '')
+	const stripped = phoneNumber.replace(PHONE_NUMBER_EXTRACT_REGEX, '')
 
-  return new RegExp(PHONE_NUMBER_VALID_REGEX).test(stripped)
+	return stripped.match(PHONE_NUMBER_VALID_REGEX) !== null
 }
 
 export default validatePhoneNumber
