+++ src/validators/rnc.validator.ts
@@ -1,6 +1,6 @@
 /*
 dr-utils: Utilities relevant to the Dominican Republic
-Copyright (C) 2021  Miguel Araujo
+Copyright (C) 2026  Miguel Araujo
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -25,37 +25,37 @@ import { validateCedula } from './cedula.validator'
  * @returns {boolean} `true` if the RNC/Cedula is valid, `false` otherwise
  */
 export function validateRNC(rnc: string): boolean {
-  const rncNoDashes = rnc.trim().replace(/-/g, '')
+	const rncNoDashes = rnc.trim().replace(/-/g, '')
 
-  if (rncNoDashes.length === 11) return validateCedula(rnc)
-  if (rncNoDashes.length !== 9) return false
+	if (rncNoDashes.length === 11) return validateCedula(rnc)
+	if (rncNoDashes.length !== 9) return false
 
-  const baseRnc = [7, 9, 8, 6, 5, 4, 3, 2]
-  let sum = 0
-  let checker
+	const baseRnc = [7, 9, 8, 6, 5, 4, 3, 2]
+	let sum = 0
+	let checker
 
-  const digits = rnc.split('').map(Number)
-  const checkDigit = digits[digits.length - 1]
+	const digits = rnc.split('').map(Number)
+	const checkDigit = digits[digits.length - 1]
 
-  for (let i = baseRnc.length - 1; i >= 0; i -= 1) {
-    sum += baseRnc[i] * digits[i]
-  }
+	for (let i = baseRnc.length - 1; i >= 0; i -= 1) {
+		sum += baseRnc[i]! * digits[i]!
+	}
 
-  const remainder = sum % 11
+	const remainder = sum % 11
 
-  switch (remainder) {
-    case 0:
-      checker = 2
-      break
-    case 1:
-      checker = 1
-      break
-    default:
-      checker = 11 - remainder
-      break
-  }
+	switch (remainder) {
+		case 0:
+			checker = 2
+			break
+		case 1:
+			checker = 1
+			break
+		default:
+			checker = 11 - remainder
+			break
+	}
 
-  return checker === checkDigit
+	return checker === checkDigit
 }
 
 export default validateRNC
