+++ tsconfig.json
@@ -1,8 +1,10 @@
 {
   "compilerOptions": {
-    "target": "ES2021",
+    "lib": ["ESNext"],
+    "target": "esnext",
     "resolveJsonModule": true,
-    "module": "commonjs",
+    "module": "preserve",
+    "moduleDetection": "force",
     "rootDir": "./src",
     "outDir": "./lib",
     "sourceMap": true,
@@ -12,7 +14,10 @@
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "skipLibCheck": true,
+    "noFallthroughCasesInSwitch": true,
+    "noUncheckedIndexedAccess": true,
+    "noImplicitOverride": true,
     "preserveSymlinks": true
   },
-  "exclude": ["node_modules", "lib"]
+  "exclude": ["node_modules", "lib", "scripts"]
 }
